requirements.txt
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
Pyrebase4==4.8.0
Requests==2.32.3
st_social_media_links==0.1.1
streamlit==1.36.0
